Further benchmarking programs
Former-commit-id: 9ad6d88d9f227092a860546fe39153266c940455 [formerly 42846f4d7f59fa01ac93446437cf38d83a9c3d40]
Former-commit-id: 96476a4ba8672d2f0b80f3817bd79abe7c2d898f

+++ src/librapid/array/functors/binary/binary.hpp
@@ -7,9 +7,9 @@ namespace librapid::functors::binary {
 		using LhsType				   = LHS;
 		using RhsType				   = RHS;
 		using RetType				   = std::false_type;
-		static constexpr int64_t Flags = 0;
+		static constexpr i64 Flags = 0;
 
-		template<typename T, i32 d, i32 a>
+		template<typename T, i64 d, i64 a>
 		LR_NODISCARD("")
 		ExtentType<T, d> genExtent(const ExtentType<T, d, a> &lhs,
 								   const ExtentType<T, d, a> &rhs) const {
@@ -17,7 +17,7 @@ namespace librapid::functors::binary {
 		}
 
 		// For a scalar operation
-		template<typename T, i32 d, i32 a>
+		template<typename T, i64 d, i64 a>
 		LR_NODISCARD("")
 		ExtentType<T, d> genExtent(const ExtentType<T, d, a> &generic) const {
 			return generic;
